% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{nyquist}
\alias{nyquist}
\title{Genera gráfico de Nyquist o Cole-Cole}
\usage{
nyquist(df, color = "red", ...)
}
\arguments{
\item{df}{Dataframe con los datos de frecuencia, Real e Imag}

\item{color}{Color de los puntos}

\item{...}{parametros usados en ggplot2}
}
\value{
Plot
}
\description{
Genera gráfico de Nyquist o Cole-Cole
}
