% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCOR.R
\name{openCOR}
\alias{openCOR}
\title{Open files of corrosion studies from different sources}
\usage{
openCOR(arch)
}
\arguments{
\item{arch}{\emph{arch} = file text from measure}
}
\value{
Dataframe(E = Potential, I = Current, time = Time)
}
\description{
Open files of corrosion studies from different sources.
This version supports \emph{.cor} and \emph{.DTA} files of the CorrWare TM software, for measuring
cyclic voltammetry, open circuit potential, and potentiodinamycs.
}
