% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCOR.R
\name{openCOR}
\alias{openCOR}
\title{\code{openCOR()}}
\usage{
openCOR(arch)
}
\arguments{
\item{arch}{\itemize{
\item *.DTA file,  Gamry Framework 7.06 version
\item *.cor file,  CorrWare TM software 2.6 and 2.6b version
#'}}
}
\value{
Dataframe con las siguentes columnas:
\itemize{
\item time: Tiempo transacurrido desde el comienzo de la
medida, (\emph{s})
\item E: Potencial, (\emph{volts}).
\item I: Corriente, (\emph{Amperes}).
\item file: Nombre del archivo.
}
}
\description{
Open files of corrosion studies from different sources.
For measuring cyclic voltammetry, open circuit potential, and potentiodinamycs.
}
\examples{
# Abre medida Voltametŕía Cíclica del software Framework
v. 7.06 de Gammry

file <- "/corrosionR/extdata/CV_Framework_version7_06.DTA"
medida <- openCOR(paste(.libPaths()[1], file, sep=""))
}
