% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openEIS.R
\name{openEIS}
\alias{openEIS}
\title{\code{openEIS()}}
\usage{
openEIS(arch)
}
\value{
dataframe con variables freq, real, imag
}
\description{
Importa datos de impedancia de distintos softwares adquisidores
}
\details{
Esta función extrae la frecuencia , parte imaginaria y parte real de los archivos generados por:

\itemize{
\item *.DTA file,  Gamry Framework 7.06 version
\item *.z file,  ZPlot software 2.6 and 2.6b version
\item *.txt ascii AUTOLAB NOVA 2.1.3 version
}
}
