% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotNyquist}
\alias{plotNyquist}
\title{Genera gráfico de Nyquist o Cole-Cole}
\usage{
plotNyquist(df, color = "red", ...)
}
\arguments{
\item{df}{Dataframe con los datos de frecuencia, Real e Imag}

\item{color}{Color de los puntos}

\item{...}{parametros usados en ggplot2}
}
\value{
plotNyquist
}
\description{
Genera gráfico de Nyquist o Cole-Cole
}
\examples{
# Abre medida de software Framework v. 7.06 de Gammry

file <- "/corrosionR/extdata/EIS_Framework_version7_06.DTA"
medida <- openEIS(paste(.libPaths()[1], file, sep=""))
plotNyquist(medida)
}
